+++ examples/convert_eeg_to_bids.py
@@ -35,7 +35,7 @@ import mne
 from mne.datasets import eegbci
 
 from mne_bids import write_raw_bids, make_bids_basename
-from mne_bids.utils import print_dir_tree
+from mne_bids.path import print_dir_tree
 
 ###############################################################################
 # Download the data
