+++ mne_bids/read.py
@@ -23,11 +23,10 @@ from mne_bids.dig import _read_dig_bids
 from mne_bids.tsv_handler import _from_tsv, _drop
 from mne_bids.config import (ALLOWED_EXTENSIONS, _convert_hand_options,
                              _convert_sex_options, reader)
-from mne_bids.utils import (_parse_bids_filename, _extract_landmarks,
-                            _find_matching_sidecar, _parse_ext,
-                            _get_ch_type_mapping, make_bids_folders,
-                            BIDSPath,
-                            _estimate_line_freq, _infer_kind)
+from mne_bids.utils import (_extract_landmarks,
+                            _get_ch_type_mapping, _estimate_line_freq)
+from mne_bids import make_bids_folders
+from mne_bids.path import BIDSPath, _parse_ext, _parse_bids_filename, _find_matching_sidecar, _infer_kind
 
 
 def _read_raw(raw_fpath, electrode=None, hsp=None, hpi=None,
